improve options doc snippets
