+++ apps/docs/src/guide/types/options.md
@@ -8,7 +8,7 @@ In this guide, we'll explain how to work with Option types in Sway and demonstra
 
 The `Option` type in Sway is a special wrapper type of Enum. In TypeScript, you can represent the `Option` type by using the `undefined` keyword, as shown in the following example
 
-<<< @/../../docs-snippets2/src/types/options.ts#options-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/types/options/options-1.ts#snippet-1{ts:line-numbers}
 
 In this example, the variable `input1` can be either a `number` or `undefined`.
 
@@ -22,14 +22,10 @@ Here's the contract function written in Sway:
 
 You can interact with the contract function using the SDK as follows:
 
-<<< @/../../docs-snippets2/src/types/options.ts#options-3{ts:line-numbers}
+<<< @/../../docs-snippets2/src/types/options/options-1.ts#snippet-2{ts:line-numbers}
 
 In this case, the result of the contract function call is the sum of both input parameters. If we pass only one parameter, the contract function will default the other parameter's value to `0`.
 
-<<< @/../../docs-snippets2/src/types/options.ts#options-4{ts:line-numbers}
+<<< @/../../docs-snippets2/src/types/options/options-2.ts#snippet-1{ts:line-numbers}
 
 Using `Option` types in Sway allows you to elegantly handle situations where a variable may or may not have a defined value.
-
-## Full Example
-
-<<< @/../../docs-snippets2/src/types/options.ts#full{ts:line-numbers}
